Move AdministrativeCriteriaOfJobApplicationForm to siae_evaluations
Sole user of that form, clarifies the scope of the form.
