+++ itou/www/eligibility_views/forms.py
@@ -71,20 +71,3 @@ class AdministrativeCriteriaForm(forms.Form):
             raise forms.ValidationError(message)
 
         return selected_objects
-
-
-class AdministrativeCriteriaOfJobApplicationForm(AdministrativeCriteriaForm):
-    def get_administrative_criteria(self):
-        return [
-            e.administrative_criteria
-            for e in self.job_application.eligibility_diagnosis.selected_administrative_criteria.select_related(
-                "administrative_criteria"
-            )
-        ]
-
-    def __init__(self, user, siae, job_application, **kwargs):
-        self.job_application = job_application
-        super().__init__(user, siae, **kwargs)
-        self.num_level2_admin_criteria = eligibilty_enums.ADMINISTRATIVE_CRITERIA_LEVEL_2_REQUIRED_FOR_SIAE_KIND[
-            siae.kind
-        ]
